requirements.txt
django==5.0.3
djangorestframework==3.14.0
mssql-django==1.41